backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pydantic==2.10.4
pydantic-settings==2.7.0
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
redis==5.2.1
httpx==0.28.1
python-dotenv==1.0.1
loguru==0.7.3
cryptography==44.0.0
